backend/reportes/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
# Base de datos SQL
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
pandas
reportlab
python-dotenv
